# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "community",
        "description": "API Documentation",
        "contact": {
            "name": "MunSeongYun",
            "email": "borygashill608@gmail.com"
        },
        "version": "0.1"
    },
    "paths": {
        "/api/comments": {
            "get": {
                "tags": [
                    "comment"
                ],
                "summary": "특정 게시글의 댓글 목록 조회",
                "description": "특정 게시글의 댓글 목록을 반환합니다.",
                "operationId": "0406f270b674fae17bf23b4206a3004f",
                "parameters": [
                    {
                        "name": "post-id",
                        "in": "query",
                        "description": "게시글 ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "댓글 목록",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 11
                                                    },
                                                    "post_id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "author": {
                                                        "type": "string",
                                                        "example": "aokon"
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "Rerum perferendis velit qui et consequatur ipsa."
                                                    },
                                                    "createdAt": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-09-30T05:45:48.000000Z"
                                                    },
                                                    "updatedAt": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-09-30T05:45:48.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "post id가 없는 경우"
                    }
                }
            },
            "post": {
                "tags": [
                    "comment"
                ],
                "summary": "새로운 댓글 생성",
                "description": "요청 본문에 포함된 데이터를 기반으로 새로운 댓글을 생성합니다.",
                "operationId": "300619a0a4003920aafe960e45688102",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "postId": {
                                        "description": "게시글 ID",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "댓글 내용",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "store successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "store successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/comments/{comment}": {
            "get": {
                "tags": [
                    "comment"
                ],
                "summary": "특정 댓글 조회",
                "description": "특정 댓글의 세부 정보를 반환합니다.",
                "operationId": "df604464ebd8be1727bb39c2711e2f03",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "댓글 ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "댓글 세부 정보",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "postId": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "Perspiciatis facilis voluptates error architecto mollitia ex sit."
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "naomi.jacobson"
                                                },
                                                "createdAt": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-09-30T05:45:46.000000Z"
                                                },
                                                "updatedAt": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-09-30T05:45:46.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "댓글을 찾을 수 없음"
                    }
                }
            },
            "put": {
                "tags": [
                    "comment"
                ],
                "summary": "특정 댓글 업데이트",
                "description": "요청 본문에 포함된 데이터를 기반으로 특정 댓글을 업데이트합니다.",
                "operationId": "11551b83dcd65f7f079adf59f8c0348e",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "댓글 ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "description": "댓글 내용",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "댓글을 찾을 수 없음"
                    }
                }
            },
            "delete": {
                "tags": [
                    "comment"
                ],
                "summary": "특정 댓글 삭제",
                "description": "특정 댓글을 삭제합니다.",
                "operationId": "e2e2b84a31c09628af8301e2ab9b209c",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "댓글 ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "댓글을 찾을 수 없음"
                    }
                }
            }
        },
...
    },
...
}